Adding "Check whether the user has a subscription or not." for seeing the detail products

+++ products/views.py
@@ -1,11 +1,14 @@
+from django.utils import timezone
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
 
+from subscriptions.models import Subscription
 from .models import Category, Product, File
 from .serializers import CategorySerializer, ProductSerializer, FileSerializer
+
 
 # class ProductListView (viewsets.ModelViewSet):
 #     queryset = Product.objects.all()
@@ -19,7 +22,7 @@ class CategoryListView(APIView):
         categories = Category.objects.all()
         serializer = CategorySerializer(categories, many=True, context={'request': request})
         return Response(data=serializer.data)
-    
+
     def post(self, request):
         serializer = CategorySerializer(data=request.data)
         if serializer.is_valid():
@@ -27,10 +30,11 @@ class CategoryListView(APIView):
             return Response(data=serializer.data, status=status.HTTP_201_CREATED)
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
 class CategoryDetailView(APIView):
 
     def get_object(self, pk):
-        try :
+        try:
             category = Category.objects.get(pk=pk)
         except Category.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -62,18 +66,20 @@ class ProductListView(APIView):
         products = Product.objects.all()
         serializer = ProductSerializer(products, many=True, context={'request': request})
         return Response(data=serializer.data)
-    
-    def post (self, request):
+
+    def post(self, request):
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(data=serializer.data, status=status.HTTP_201_CREATED)
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
 class ProductDetailView(APIView):
     permission_classes = [IsAuthenticated]
 
     def get_object(self, pk):
+
         try:
             product = Product.objects.get(pk=pk)
         except Product.DoesNotExist:
@@ -81,11 +87,23 @@ class ProductDetailView(APIView):
         return product
 
     def get(self, request, pk):
+        if not Subscription.objects.filter(
+            user=request.user,
+            expire_time__gt=timezone.now()
+        ).exists():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         product = self.get_object(pk)
         serializer = ProductSerializer(product, context={'request': request})
         return Response(data=serializer.data)
-    
-    def put (self, request, pk):
+
+    def put(self, request, pk):
+        if not Subscription.objects.filter(
+            user=request.user,
+            expire_time__gt=timezone.now()
+        ).exists():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         product = self.get_object(pk)
         serializer = ProductSerializer(product, data=request.data, context={'request': request})
         if serializer.is_valid():
@@ -94,9 +112,16 @@ class ProductDetailView(APIView):
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        if not Subscription.objects.filter(
+            user=request.user,
+            expire_time__gt=timezone.now()
+        ).exists():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         product = self.get_object(pk)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
 
 class FileListView(APIView):
 
@@ -104,13 +129,14 @@ class FileListView(APIView):
         files = File.objects.filter(product_id=product_id)
         serializer = FileSerializer(files, many=True, context={'request': request})
         return Response(data=serializer.data)
-    
+
     def post(self, request, product_id):
         serializer = FileSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save(product_id=product_id)
             return Response(data=serializer.data, status=status.HTTP_201_CREATED)
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 class FileDetailView(APIView):
 
@@ -121,20 +147,20 @@ class FileDetailView(APIView):
             return Response(status=status.HTTP_404_NOT_FOUND)
         return file
 
-    def get(self, request, product_id , pk):
+    def get(self, request, product_id, pk):
         file = self.get_object(pk)
         serializer = FileSerializer(file, context={'request': request})
         return Response(data=serializer.data)
-    
-    def put(self, request, product_id , pk):
+
+    def put(self, request, product_id, pk):
         file = self.get_object(product_id, pk)
         serializer = FileSerializer(file, data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save()
             return Response(data=serializer.data, status=status.HTTP_201_CREATED)
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    def delete(self, request, product_id , pk):
+
+    def delete(self, request, product_id, pk):
         file = self.get_object(product_id, pk)
         file.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
